Async territory updates. Works pretty well, can fill an entire level with no lag spikes. There is currently a very slight level of degradation with a completely packed level full of actors (unrealistic for gameplay) but that is probably better in a cooked build. I'll make sure later that actors arent ticking/replicating unless they need to.
Committer's note: pushing the current state in case a world-shattering earthquake consumes my computer or something. Will be out of band to work for a while. Maybe 4 more years, who knows. By the way, I think some new NodeGenerator stuff is making it in inadvertently. Next step is to rework that into code as I borked the blueprint with recent code migrations.